// src/types/apps/account.ts
export interface UserType {
  user_id: number | null;
  user_type: number;
  user_email: string; 
  user_password: string;
  register_num: string;
  company_name: string;
  company_address: string;
  manager_name: string;
  manager_phone: string;
  manager_depart: string;
  manager_grade: string;
  other: string;
  admin_name: string;
  admin_phone: string;
  approval: number;
  id: string;
}
 